Update LOOT `dirty` masterlist entry template
All masterlists except Oblivion's use the `dirtyPlugin` YAML anchor,
which was previously absent from the output. The masterlists also are
moving away from using the `dirtyUtil` anchor instead writing utility
name and version for each record directly. (The `dirtyPlugin` now has
the information that used to be contained in the `dirtyUtil` anchor.)

Also removes lowercasing of the CRC hex. All tools, including the xEdit
and LOOT interfaces, output with uppercase hexadecimals and the
masterlist itself is case insensitive. Using uppercase makes the
hexadecimal CRCs appear the same across the board.