remember previous selection in the "Copy as [X] into" module selection dialog.
when opening the module selection dialog puts the focus on the first checked node and selects all checked nodes.

If only a single node is focused/selected/checked at this time, the check mark is removed (as simply hitting enter will check that single node again anyway, but unchecking it allows to then simply use the cursor keys to move to a different node and hit enter without first having to uncheck the previously checked node).