allow collapsing nodes in view tab
collapsing a single node is stored in the elements and applies only for that one record. It may also get forgotten when navigating away from the record and coming back later (depends if the element gets freed or not, which depends on exactly which element it is and if it has been modified or not).

When holding shift while clicking the [-] on front of the node, the collapse is stored in the Def instead of the Element and applies to all Elements of that definition, so will be preserved when navigating between different records.

When expanding a node that was globally collapsed (by holding shift), it will be globally expanded.